flashcards and keywords can both be queried for one and all

+++ db/migrations/migration-08292018.sql
@@ -3,7 +3,7 @@ CREATE TABLE IF NOT EXISTS flashcards (
   question TEXT,
   answer TEXT,
   category VARCHAR(255),
-  difficulty BIGINT
+  difficulty INTEGER
 );
 
 CREATE INDEX idx_category ON flashcards (category);
